feat: add PHP 8.2 support (dvsa/olcs-backend#172)
* feat!: php8.2 support initial work

* feat!: PHP ~8.2.0 support initial work

* feat!: PHP 8.2

* feat!: PHP 8.2 work

* feat!: PHP 8.2 work

* feat!: PHP 8.2 work

* feat!: PHP 8.2 work

* feat!: PHP 8.2 work

* feat: empty to trigger CI

* fix: fix remaining CI and tests, standardise how we bring back config (lowercase c)

* chore: remove rector following php 8.2 update

---------

+++ app/api/module/DocumentShare/src/Data/Object/File.php
@@ -42,7 +42,7 @@ class File
      */
     public function __destruct()
     {
-        if (is_file($this->file)) {
+        if ($this->file !== null && is_file($this->file)) {
             unlink($this->file);
         }
     }
